Range Voting store implemented
