+++ apps/range-voting/app/store/app.js
@@ -1,0 +1,2 @@
+import Aragon from '@aragon/api'
+export const app = new Aragon()
