+++ apps/range-voting/app/store/index.js
@@ -1,0 +1,4 @@
+export { app } from './app'
+export { handleEvent } from './events'
+export { initStore } from './init'
+export const INITIALIZATION_TRIGGER = Symbol('INITIALIZATION_TRIGGER')
