+++ apps/range-voting/app/store/rxjs.js
@@ -1,0 +1,3 @@
+import 'rxjs/add/operator/first'
+
+export { of } from 'rxjs/observable/of'
